% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coffee_cupping_tasting}
\alias{coffee_cupping_tasting}
\title{Coffee cupping flavors.}
\format{
A data frame with 110 rows and 6 variables:
\describe{
 \item{ids}{final label}
  \item{end_name}{labels for id}
  \item{parents}{first tier name}
  \item{2nd level}{second tier name}
  \item{labels_to_attach}{used to create tier names}
  \item{labels}{hover text}
}
}
\source{
\url{https://worldcoffeeresearch.org/work/sensory-lexicon/}
}
\usage{
coffee_cupping_tasting
}
\description{
A lexicon of flavors.
}
\keyword{datasets}
