% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coffee_flavors}
\alias{coffee_flavors}
\title{Coffee cupping flavors.}
\format{
A data frame with 110 rows and 6 variables
}
\usage{
coffee_flavors
}
\description{
Coffee cupping flavors.
}
\keyword{datasets}
