% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_prepare_csv.R
\name{first_line_of_csv}
\alias{first_line_of_csv}
\title{Get first line for csv}
\usage{
first_line_of_csv(wide_format)
}
\arguments{
\item{wide_format}{profile as df}
}
\value{
string of first line
}
\description{
Get first line for csv
}
