% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_extract_from_json.R
\name{get_data_of_phase_times}
\alias{get_data_of_phase_times}
\title{Get 3 phase lengths}
\usage{
get_data_of_phase_times(json_file)
}
\arguments{
\item{json_file}{json file that is read}
}
\description{
Get 3 phase lengths
}
