% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_extract_from_json.R
\name{get_data_of_times_temps}
\alias{get_data_of_times_temps}
\title{Get times and temps for graph. Subract TP_idx from all times}
\usage{
get_data_of_times_temps(json_file)
}
\arguments{
\item{json_file}{json file that is read}
}
\description{
It also subracts TP_idx from all times so start time = charge time.
}
