% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_extract_from_json.R
\name{get_data_to_display_at_upload}
\alias{get_data_to_display_at_upload}
\title{Display data at upload # CHECK IF NO DROP TIME check}
\usage{
get_data_to_display_at_upload(json_file)
}
\arguments{
\item{json_file}{json file that is read}
}
\value{
df in wide format
}
\description{
Grabs certain fields for display.
}
