% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_extract_from_json.R
\name{get_event_times}
\alias{get_event_times}
\title{Get tp, dry, fc, sc, drop times}
\usage{
get_event_times(json_file)
}
\arguments{
\item{json_file}{json file that is read}
}
\description{
Get tp, dry, fc, sc, drop times
}
