% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{haiti}
\alias{haiti}
\title{Coffee cupping flavors.}
\format{
A data frame with 110 rows and 6 variables:
\describe{
  \item{end_name}{name of taste}
  \item{labels}{description of tastes}
  ...
}
}
\source{
\url{https://worldcoffeeresearch.org/work/sensory-lexicon/}
}
\usage{
haiti
}
\description{
A dataset containing a lexicon of coffee tastes broken.
}
\keyword{datasets}
